Sort the output of `GET /api/v1/vaas` by timestamp
The output of `GET /api/v1/vaas` was not being correctly sorted.
This commit changes the behavior so that the output will be sorted by
the `timestamp` field.

Additionally, duplicated code was removed in the VAA code that handles
queries.

Also, there was an unused query parameter `sortBy` which was
being exposed but not used. Hence it was removed.