Add log messages on API startup
Often times, the API service fails to start due to connectivity
issues with other services (e.g.: redis).

Previously, no log information was being emitted, which made it hard
to troubleshoot the exact reason of the service failing to start.

This commit adds log messages when the service starts to make
those cases easier to debug.